+++ src/Yajima/y_runtime.c
@@ -5305,7 +5305,7 @@ void FUN_00434d30(void)
 
 {
   char cVar1;
-  u8 uVar2;
+  int uVar2;
   int iVar3;
   
   for (iVar3 = 1; iVar3 < 4; iVar3 = iVar3 + 1) {
